backend/requirements.txt
Django==5.0.1
typing_extensions==4.9.0
channels[daphne]==4.0.0
django-cors-headers==4.3.1
psycopg[binary]==3.1.17
django-ninja==1.1.0
requests==2.31.0
pyjwt==2.8.0
Pillow==10.2.0